requirements.txt
transformers
openai-whisper
ffmpeg-python
